refactor: adjust namespaces that are confusing to stick to process names
