+++ Chapter-1-initial-architecture/Src/Fitnet.ArchitectureTests/OffersArchitectureTests.cs
@@ -38,7 +38,7 @@ public sealed class OffersArchitectureTests
         var offersModule = Solution.Types
             .That()
             .ResideInNamespace(Modules.Offers);
-
+ 
         var shouldModule = Solution.Types
             .That()
             .ResideInNamespace(Modules.Passes)
