--- Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Api/Prepare/PrepareContractRequest.cs
@@ -1,8 +1,0 @@
-namespace EvolutionaryArchitecture.Fitnet.Contracts.Api.Prepare;
-
-using Application.Prepare;
-
-internal sealed record PrepareContractRequest(Guid CustomerId, int CustomerAge, int CustomerHeight, DateTimeOffset PreparedAt)
-{
-    internal PrepareContractCommand ToCommand() => new(CustomerId, CustomerAge, CustomerHeight, PreparedAt);
-}
