+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Api/PrepareContract/PrepareContractRequest.cs
@@ -1,0 +1,8 @@
+namespace EvolutionaryArchitecture.Fitnet.Contracts.Api.PrepareContract;
+
+using EvolutionaryArchitecture.Fitnet.Contracts.Application.PrepareContract;
+
+internal sealed record PrepareContractRequest(Guid CustomerId, int CustomerAge, int CustomerHeight, DateTimeOffset PreparedAt)
+{
+    internal PrepareContractCommand ToCommand() => new(CustomerId, CustomerAge, CustomerHeight, PreparedAt);
+}
