+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Api/PrepareContract/PrepareContractRequestValidator.cs
@@ -1,0 +1,14 @@
+namespace EvolutionaryArchitecture.Fitnet.Contracts.Api.PrepareContract;
+
+using FluentValidation;
+
+internal sealed class PrepareContractRequestValidator : AbstractValidator<PrepareContractRequest>
+{
+    public PrepareContractRequestValidator()
+    {
+        RuleFor(request => request.CustomerId).NotEmpty();
+        RuleFor(request => request.CustomerAge).GreaterThan(0);
+        RuleFor(request => request.CustomerHeight).GreaterThan(0);
+        RuleFor(request => request.PreparedAt).NotEmpty();
+    }
+}
