--- Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Api/Sign/SignContractRequest.cs
@@ -1,9 +1,0 @@
-namespace EvolutionaryArchitecture.Fitnet.Contracts.Api.Sign;
-
-using Application.Sign;
-
-internal sealed record SignContractRequest(DateTimeOffset SignedAt)
-{
-    internal SignContractCommand ToCommand(Guid id) =>
-        new(id, SignedAt);
-}
