--- Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Api/Sign/SignContractRequestValidator.cs
@@ -1,9 +1,0 @@
-namespace EvolutionaryArchitecture.Fitnet.Contracts.Api.Sign;
-
-using FluentValidation;
-
-internal sealed class SignContractRequestValidator : AbstractValidator<SignContractRequest>
-{
-    public SignContractRequestValidator() => RuleFor(signContractRequest => signContractRequest.SignedAt)
-        .NotEmpty();
-}
