+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Api/SignContract/SignContractRequest.cs
@@ -1,0 +1,9 @@
+namespace EvolutionaryArchitecture.Fitnet.Contracts.Api.SignContract;
+
+using EvolutionaryArchitecture.Fitnet.Contracts.Application.SignContract;
+
+internal sealed record SignContractRequest(DateTimeOffset SignedAt)
+{
+    internal SignContractCommand ToCommand(Guid id) =>
+        new(id, SignedAt);
+}
