+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Api/SignContract/SignContractRequestValidator.cs
@@ -1,0 +1,9 @@
+namespace EvolutionaryArchitecture.Fitnet.Contracts.Api.SignContract;
+
+using FluentValidation;
+
+internal sealed class SignContractRequestValidator : AbstractValidator<SignContractRequest>
+{
+    public SignContractRequestValidator() => RuleFor(signContractRequest => signContractRequest.SignedAt)
+        .NotEmpty();
+}
