--- Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Application/Prepare/PrepareContractCommand.cs
@@ -1,3 +1,0 @@
-namespace EvolutionaryArchitecture.Fitnet.Contracts.Application.Prepare;
-
-public sealed record PrepareContractCommand(Guid CustomerId, int CustomerAge, int CustomerHeight, DateTimeOffset PreparedAt) : ICommand<Guid>;
