+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Application/PrepareContract/PrepareContractCommand.cs
@@ -1,0 +1,3 @@
+namespace EvolutionaryArchitecture.Fitnet.Contracts.Application.PrepareContract;
+
+public sealed record PrepareContractCommand(Guid CustomerId, int CustomerAge, int CustomerHeight, DateTimeOffset PreparedAt) : ICommand<Guid>;
