+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Application/PrepareContract/PrepareContractCommandHandler.cs
@@ -1,0 +1,16 @@
+namespace EvolutionaryArchitecture.Fitnet.Contracts.Application.PrepareContract;
+
+using Core;
+
+[UsedImplicitly]
+internal sealed class PrepareContractCommandHandler(IContractsRepository contractsRepository) : IRequestHandler<PrepareContractCommand, Guid>
+{
+    public async Task<Guid> Handle(PrepareContractCommand command, CancellationToken cancellationToken)
+    {
+        var previousContract = await contractsRepository.GetPreviousForCustomerAsync(command.CustomerId, cancellationToken);
+        var contract = Contract.Prepare(command.CustomerId, command.CustomerAge, command.CustomerHeight, command.PreparedAt, previousContract?.IsSigned);
+        await contractsRepository.AddAsync(contract, cancellationToken);
+
+        return contract.Id.Value;
+    }
+}
