--- Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Application/Sign/SignContractCommand.cs
@@ -1,3 +1,0 @@
-namespace EvolutionaryArchitecture.Fitnet.Contracts.Application.Sign;
-
-public sealed record SignContractCommand(Guid Id, DateTimeOffset SignedAt) : ICommand;
