+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Application/SignContract/SignContractCommand.cs
@@ -1,0 +1,3 @@
+namespace EvolutionaryArchitecture.Fitnet.Contracts.Application.SignContract;
+
+public sealed record SignContractCommand(Guid Id, DateTimeOffset SignedAt) : ICommand;
