--- Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Application/Terminate/TerminateBindingContractCommand.cs
@@ -1,3 +1,0 @@
-namespace EvolutionaryArchitecture.Fitnet.Contracts.Application.Terminate;
-
-public sealed record TerminateBindingContractCommand(Guid Id) : ICommand;
