--- Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Application/Terminate/TerminateBindingContractCommandHandler.cs
@@ -1,19 +1,0 @@
-namespace EvolutionaryArchitecture.Fitnet.Contracts.Application.Terminate;
-
-using Common.Api.ErrorHandling;
-using Core;
-
-[UsedImplicitly]
-internal sealed class TerminateBindingContractCommandHandler(
-    IBindingContractsRepository bindingContractsRepository,
-    TimeProvider timeProvider) : IRequestHandler<TerminateBindingContractCommand>
-{
-    public async Task Handle(TerminateBindingContractCommand command, CancellationToken cancellationToken)
-    {
-        var contract = await bindingContractsRepository.GetByIdAsync(command.Id, cancellationToken) ??
-                       throw new ResourceNotFoundException(command.Id);
-        var now = timeProvider.GetUtcNow();
-        contract.Terminate(now);
-        await bindingContractsRepository.CommitAsync(cancellationToken);
-    }
-}
