+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Application/TerminateBindingContract/TerminateBindingContractCommand.cs
@@ -1,0 +1,3 @@
+namespace EvolutionaryArchitecture.Fitnet.Contracts.Application.TerminateBindingContract;
+
+public sealed record TerminateBindingContractCommand(Guid Id) : ICommand;
