+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Application/TerminateBindingContract/TerminateBindingContractCommandHandler.cs
@@ -1,0 +1,19 @@
+namespace EvolutionaryArchitecture.Fitnet.Contracts.Application.TerminateBindingContract;
+
+using EvolutionaryArchitecture.Fitnet.Common.Api.ErrorHandling;
+using EvolutionaryArchitecture.Fitnet.Contracts.Core;
+
+[UsedImplicitly]
+internal sealed class TerminateBindingContractCommandHandler(
+    IBindingContractsRepository bindingContractsRepository,
+    TimeProvider timeProvider) : IRequestHandler<TerminateBindingContractCommand>
+{
+    public async Task Handle(TerminateBindingContractCommand command, CancellationToken cancellationToken)
+    {
+        var contract = await bindingContractsRepository.GetByIdAsync(command.Id, cancellationToken) ??
+                       throw new ResourceNotFoundException(command.Id);
+        var now = timeProvider.GetUtcNow();
+        contract.Terminate(now);
+        await bindingContractsRepository.CommitAsync(cancellationToken);
+    }
+}
