+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Core/BindingContract.cs
@@ -3,8 +3,8 @@
 using Common.Core.BusinessRules;
 using DomainDrivenDesign.BuildingBlocks;
 using SignContract;
-using TerminateContract;
-using TerminateContract.BusinessRules;
+using TerminateBindingContract;
+using TerminateBindingContract.BusinessRules;
 
 public sealed class BindingContract : Entity
 {
