+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Core/TerminateBindingContract/BindingContractTerminatedEvent.cs
@@ -1,0 +1,12 @@
+﻿namespace EvolutionaryArchitecture.Fitnet.Contracts.Core.TerminateBindingContract;
+
+using EvolutionaryArchitecture.Fitnet.DomainDrivenDesign.BuildingBlocks;
+
+public sealed record BindingContractTerminatedEvent(
+    Guid Id,
+    DateTimeOffset TerminatedAt,
+    DateTime OccuredAt) : IDomainEvent
+{
+    internal static BindingContractTerminatedEvent Raise(DateTimeOffset terminatedAt)
+        => new(Guid.NewGuid(), terminatedAt, DateTime.Now);
+}
