+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.IntegrationTests/PrepareContract/PrepareContractTests.cs
@@ -3,7 +3,7 @@ namespace EvolutionaryArchitecture.Fitnet.Contracts.IntegrationTests.PrepareCont
 using EvolutionaryArchitecture.Fitnet.Common.IntegrationTestsToolbox.TestEngine.Database;
 using EvolutionaryArchitecture.Fitnet.Common.Api.ErrorHandling;
 using Api;
-using Api.Prepare;
+using Api.PrepareContract;
 using Fitnet.Common.IntegrationTestsToolbox.TestEngine;
 using Fitnet.Common.IntegrationTestsToolbox.TestEngine.Configuration;
 using Fitnet.Common.IntegrationTestsToolbox.TestEngine.EventBus;
