+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.IntegrationTests/SignContract/SignContractTestExtensions.cs
@@ -1,6 +1,6 @@
 ﻿namespace EvolutionaryArchitecture.Fitnet.Contracts.IntegrationTests.SignContract;
 
-using Api.Sign;
+using Api.SignContract;
 
 internal static class SignContractTestExtensions
 {
